Деление теперь нецелочисленное

+++ main.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main() {
-	int number;
+	double number;
 	cout << "Hello, dearest User!" << endl;
 	cout << "Think number... and input it here " << endl;
 	cin >> number;
@@ -17,7 +17,7 @@ int main() {
 	
 	cout << "The programm works OK..." << endl;
 
-	int another_number;
+	double another_number;
 	cout << "Input another nember here " << endl;
 	cin >> another_number;
 	while (number == another_number) {
